+++ test/run_fail/test_assert_equal.js
