+++ test/run_fail/test_events_uncaught_error.js
