+++ test/run_fail/test_iotjs_syntax_error.js
