+++ test/run_fail/test_module_require_invalid_file.js
