+++ test/run_fail/test_process_explicit_exit.js
