+++ test/run_fail/test_process_implicit_exit.js
