+++ test/run_pass/test_events_emit_error.js
