+++ test/run_pass/test_events_uncaught_error.js
