+++ test/run_pass/test_fs_event.js
