+++ test/run_pass/test_fs_exists_sync.js
