+++ test/run_pass/test_fs_open_read.js
@@ -25,17 +25,6 @@ var flags = "r";
 var mode = 438;
 
 
-// test sync open & read
-try {
-  var fd = fs.openSync(fileName, flags, mode);
-  var buffer = new Buffer(64);
-  fs.readSync(fd, buffer, 0, buffer.length, 0);
-  assert.equal(buffer.toString(), expectedContents);
-} catch (err) {
-  throw err;
-}
-
-
 // test async open & read
 fs.open(fileName, flags, mode, function(err, fd) {
   if (err) {
