+++ test/run_pass/test_fs_open_read_sync_1.js
