+++ test/run_pass/test_fs_readfile_sync.js
