+++ test/run_pass/test_gpio_input.js
