+++ test/run_pass/test_gpio_output.js
