+++ test/run_pass/test_net_1.js
