+++ test/run_pass/test_net_10.js
