+++ test/run_pass/test_net_2.js
