+++ test/run_pass/test_net_3.js
