+++ test/run_pass/test_net_4.js
