+++ test/run_pass/test_net_5.js
