+++ test/run_pass/test_net_6.js
