+++ test/run_pass/test_net_7.js
