+++ test/run_pass/test_net_8.js
