+++ test/run_pass/test_net_9.js
