+++ test/run_pass/test_net_http_get.js
