+++ test/run_pass/test_process_cwd.js
