+++ test/run_pass/test_process_exit.js
