+++ test/run_pass/test_process_experimental_on.js
