+++ test/run_pass/test_process_next_tick.js
