+++ test/run_pass/test_process_uncaught_order.js
