+++ test/run_pass/test_process_uncaught_simple.js
