+++ test/run_pass/test_timers_arguments.js
