+++ test/run_pass/test_timers_simple.js
